# Do not allocate a new map when the value is the same