@garazdawi 2018-07-31
Looks good, putting in tests.